% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{alphash}
\alias{alphash}
\title{Create "hashes" of the letters in a character vector}
\usage{
alphash(x, case_sensitive = FALSE, alphabet = standard_alphabet)
}
\arguments{
\item{x}{A character vector to be hashed}

\item{case_sensitive}{NOT IN USE. This parameter will probably be removed}

\item{alphabet}{NOT IN USE. This parameter will probably be removed}
}
\value{
A character vector equal in length to the input x
}
\description{
The "hash" here is not a true hash, since the output has a variable number of
characters, but is instead a string representation of the number of characters
in the input string in hexadecimal format.
}
