% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{analyse_frequency}
\alias{analyse_frequency}
\title{Make a frequency table of the characters in a string}
\usage{
analyse_frequency(
  x,
  case_sensitive = FALSE,
  alphabet = standard_alphabet,
  collapse = FALSE
)
}
\arguments{
\item{x}{A character vector}

\item{case_sensitive}{Logical. Should lower and upper case letters be kept separate}

\item{alphabet}{A list containing lower and upper case alphabets}
}
\value{
A tibble of the frequencies, with one row for each input string
}
\description{
Make a frequency table of the characters in a string
}
