% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\docType{data}
\name{base64}
\alias{base64}
\title{A helper table for base64 conversion and looup}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 3 columns.}
\usage{
base64
}
\description{
A helper table for base64 conversion and looup
}
\keyword{datasets}
