% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_gc}
\alias{convert_gc}
\title{Convert between GC codes and GC ids}
\usage{
convert_gc(x)
}
\arguments{
\item{x}{either a valid GC code, or the id of the cache}
}
\value{
a list of conversions, each element being a list of length two.
The first element will be called \code{code} and will be the GC code.
The second element will be \code{id} and will be a character value of the equivalent id
}
\description{
Convert between GC codes and GC ids
}
\examples{
convert_gc("12345")
convert_gc(54321)
convert_gc("GC3NWRQ")
}
