% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_latlon_conversion.R
\name{degrees}
\alias{degrees}
\title{Convert radians to degrees}
\usage{
degrees(x)
}
\arguments{
\item{x}{A numeric vector of radians}
}
\value{
A numeric vector of degrees
}
\description{
Convert radians to degrees
}
\examples{
degrees(0.9696273)
}
