% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{expressCoordinates}
\alias{expressCoordinates}
\title{Express Decimal Coordinates in Other (text) Formats}
\usage{
expressCoordinates(x, style = "GC")
}
\arguments{
\item{x}{A numeric vector of length 2}

\item{style}{placeholder for future development if requirements emerge}
}
\value{
A character of length 1 with an alternative expression of the coordinates
}
\description{
Designed to convert into Geocaching-style style coordinates, but future styles may be accommodated.
}
\examples{
expressCoordinates(c(55.9327, -3.25103))

}
