% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_latlon_conversion.R
\name{ll_to_os}
\alias{ll_to_os}
\title{Convert from latitude & longitude to British Grid}
\usage{
ll_to_os(phi, lam, datum = "osgb36")
}
\arguments{
\item{phi}{The latitude as decimal degrees}

\item{lam}{The longitude as decimal degrees}

\item{datum}{A character string naming the datum to convert to}
}
\value{
A numeric vector of two: the East and North of the grid coordinates
}
\description{
Convert from latitude & longitude to British Grid
}
\details{
only datum = "osgb36" currently supported
}
\examples{
ll_to_os()
}
