% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_latlon_conversion.R
\name{os_to_ll}
\alias{os_to_ll}
\title{Convert from British Grid to latitude & longitude}
\usage{
os_to_ll(E, N, datum = "osgb36")
}
\arguments{
\item{E}{The East part of the grid coordinates}

\item{N}{The North part of the grid coordinates}

\item{datum}{A character string naming the datum to convert from}
}
\value{
A numeric vector of two: the latitude and longitude
}
\description{
Convert from British Grid to latitude & longitude
}
\details{
only datum = "osgb36" currently supported
}
\examples{
os_to_ll(337730, 679013)
}
