% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{parseCoordinates}
\alias{parseCoordinates}
\title{importFrom magrittr `%>%`
importFrom magrittr `%<>%`
Parse Coordinates into Numeric Format}
\usage{
parseCoordinates(x)
}
\arguments{
\item{x}{A string for the coordinates to be converted}
}
\value{
A numeric vector holding the n(orth) and e(ast) coordinates
}
\description{
parseCoordinates takes a variety of string inputs for coordinates in the
following formats:
- N00 00.000 W000 00.000
- N00 00 00 W000 00 00
- N00.0000 W00.0000
and converts them into a numeric vector of length 2
}
\examples{
parseCoordinates("N55 55.555 W003 14.159")
parseCoordinates("N 55 55.555   E003 14.159")
parseCoordinates("N55.92592 W3.23598")
}
