% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{qqmiaiii}
\alias{qqmiaiii}
\title{Encrypt a string using the Vigenere cipher}
\usage{
qqmiaiii(x, key, alphabet = standard_alphabet)
}
\arguments{
\item{x}{A string to encrypt or decrypt}

\item{key}{The encryption or decryption key}

\item{alphabet}{A list of letters in lower and upper case}
}
\description{
This is a wrapper for \code{vigenere} where decrypt is set to FALSE
}
\seealso{
\code{\link{vigenere}}
}
