% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_latlon_conversion.R
\name{radians}
\alias{radians}
\title{Convert degrees to radians}
\usage{
radians(x)
}
\arguments{
\item{x}{A numeric vector of degrees}
}
\value{
A numeric vector of radians
}
\description{
Convert degrees to radians
}
\examples{
radians(55.55555)
}
