% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{rot}
\alias{rot}
\title{rot a string by a given number of letters.}
\usage{
rot(x, n = 13, alphabet = list(lw = letters, up = LETTERS))
}
\arguments{
\item{x}{A string}

\item{n}{A number of letters to rot the string by}

\item{alphabet}{A list containing lower and upper case alphabets}
}
\value{
A string
}
\description{
rot a string by a given number of letters.
}
\examples{
rot("abc")
rot("abc", n=2)
rot("abc", n=5, list(lw=letters[1:7], up=LETTERS[1:7]))
}
