% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{rot_all}
\alias{rot_all}
\title{rot a string over all possible n}
\usage{
rot_all(x, alphabet = standard_alphabet)
}
\arguments{
\item{x}{A string.}

\item{alphabet}{A list containing lower and upper case alphabets.}
}
\value{
a vector of strings
}
\description{
rot a string over all possible n
}
\examples{
rot_all("abc")
rot_all("abc", list(lw=letters[1:7], up=LETTERS[1:7]))
}
