% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\docType{data}
\name{Scrabble}
\alias{Scrabble}
\title{Value and frequency of Scrabble letters}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 3 columns.
}
\usage{
Scrabble
}
\description{
Value and frequency of Scrabble letters
}
\keyword{datasets}
