% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{Scrabble_score}
\alias{Scrabble_score}
\title{Find the Scrabble value of words}
\usage{
Scrabble_score(x, language = "en")
}
\arguments{
\item{x}{A vector of character strings}

\item{language}{A character string for the linguistic Scrabble edition, conforming to ISO 639-1

Current supported languages: en}
}
\value{
An integer vector
}
\description{
Find the Scrabble value of words
}
