% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.R
\name{spiralise}
\alias{spiralise}
\title{Populate a vector into a matrix in a spiral patter}
\usage{
spiralise(x, r, c)
}
\arguments{
\item{x}{A vector to spiralise}

\item{r}{An integer; the number of rows}

\item{c}{An integer; the number of columns}
}
\value{
A matrix with r rows and c columns
}
\description{
Populate a vector into a matrix in a spiral patter
}
\examples{
spiralise(1:48, 4, 12)

}
