% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\docType{data}
\name{standard_alphabet}
\alias{standard_alphabet}
\title{The standard alphabet for the locale, for use in Caeser-based encryption etc.}
\format{An object of class \code{list} of length 2.}
\usage{
standard_alphabet
}
\description{
The standard alphabet for the locale, for use in Caeser-based encryption etc.
}
\keyword{datasets}
