% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{tvaersum}
\alias{tvaersum}
\alias{digital_root}
\title{Finds the digital root (tvaersum in Danish) of a number}
\usage{
tvaersum(x, warnDecimal = TRUE)
}
\arguments{
\item{x}{A number}

\item{warnDecimal}{Logical. If TRUE, a warning will be issued if x contains a decimal}
}
\value{
An integer
}
\description{
Finds the digital root (tvaersum in Danish) of a number
}
\examples{
tvaersum(91256015)
tvaersum("(844) 448-1212)")

}
