% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{w3w}
\alias{w3w}
\title{What 3 Words wrapper}
\usage{
w3w(x)
}
\arguments{
\item{x}{Either a vector of words, for a single latitude/longitude pair, or
a list of vectors for vectorised operations. This wrapper also accepts a
single string of three words separated by full stops.}
}
\value{
a thing
}
\description{
What 3 Words wrapper
}
\examples{
blah()
}
