% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{word_score}
\alias{word_score}
\title{Find the value of words}
\usage{
word_score(x)
}
\arguments{
\item{x}{A vector of character strings}
}
\value{
An integer vector
}
\description{
Find the value of words
}
\examples{
word_score(c("infinite", "monkey", "cage"))

}
