% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{expressCoordinates}
\alias{expressCoordinates}
\title{Express decimal coordinates in other (text) formats}
\usage{
expressCoordinates(x, style = "GC")
}
\description{
Express decimal coordinates in other (text) formats
}
