% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{rot}
\alias{rot}
\title{Caesar-shift a string by a given number of letters.}
\usage{
rot(x, n = 13, alphabet = standard_alphabet, showWarn = TRUE)
}
\arguments{
\item{x}{A string.}

\item{n}{A number of letters to shift the string by.}

\item{alphabet}{A list containing lower and upper case alphabets}

\item{showWarn}{boolean. Do you want to see warnings about alphabets?}
}
\value{
A string
}
\description{
Caesar-shift a string by a given number of letters.
}
\examples{
rot("abc")
rot("abc", n=2)
rot("abc", n=5, list(lw=letters[1:7], up=LETTERS[1:7]))
}
