% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{vigenere}
\alias{vigenere}
\title{Encrypt or decrypt a string using a key}
\usage{
vigenere(x, key, decrypt = TRUE, alphabet = standard_alphabet)
}
\arguments{
\item{x}{A string to encrypt or decrypt}

\item{key}{The encryption or decryption key}

\item{decrypt}{Are you decrypting an encrypted string?}

\item{alphabet}{A list of letters in lower and upper case}
}
\value{
A string
}
\description{
Encrypt or decrypt a string using a key
}
\examples{
vigenere("MN vdopf wq brcep zwtcd.", "midway")
vigenere("My treasure is buried he... find it who may.", "La Bouche", decrypt = FALSE)

}
